refactor: 🎸 move Ray orchestration into Ray config
